Move to de.westnordost.streetcomplete.data.meta.OsmTaggings
